+++ src/main/java/com/gmail/nossr50/commands/skills/HerbalismCommand.java
@@ -11,8 +11,8 @@ import com.gmail.nossr50.datatypes.SkillType;
 import com.gmail.nossr50.locale.LocaleLoader;
 
 public class HerbalismCommand extends SkillCommand {
-	AdvancedConfig advancedConfig = AdvancedConfig.getInstance();
-	
+    AdvancedConfig advancedConfig = AdvancedConfig.getInstance();
+    
     private String greenTerraLength;
     private String greenThumbChance;
     private String greenThumbStage;
@@ -28,7 +28,7 @@ public class HerbalismCommand extends SkillCommand {
     private int greenThumbMaxLevel = advancedConfig.getGreenThumbMaxLevel();
     private double doubleDropsMaxBonus = advancedConfig.getHerbalismDoubleDropsChanceMax();
     private int doubleDropsMaxLevel = advancedConfig.getHerbalismDoubleDropsMaxLevel();
-	
+    
     private boolean canGreenTerra;
     private boolean canGreenThumbWheat;
     private boolean canGreenThumbBlocks;
@@ -42,21 +42,21 @@ public class HerbalismCommand extends SkillCommand {
 
     @Override
     protected void dataCalculations() {
-    	DecimalFormat df = new DecimalFormat("#.0");
-    	greenTerraLength = String.valueOf(2 + ((int) skillValue / abilityLengthIncreaseLevel));
-    	//FARMERS DIET
-    	if(skillValue >= farmersDietMaxLevel) farmersDietRank = "5";
-    	else farmersDietRank = String.valueOf((int)skillValue / farmersDietRankChange);
-    	//GREEN THUMB
-    	if(skillValue >= greenThumbStageMaxLevel) greenThumbStage = "4";
-    	else greenThumbStage = String.valueOf((int)skillValue / greenThumbStageChange);
-    	
-    	
-    	if(skillValue >= greenThumbMaxLevel) greenThumbChance = String.valueOf(greenThumbMaxBonus);
-    	else greenThumbChance = String.valueOf((greenThumbMaxBonus / greenThumbMaxLevel) * skillValue);
-    	//DOUBLE DROPS
-    	if(skillValue >= doubleDropsMaxLevel) doubleDropChance = df.format(doubleDropsMaxBonus);
-    	else doubleDropChance = df.format((doubleDropsMaxBonus / doubleDropsMaxLevel) * skillValue);
+        DecimalFormat df = new DecimalFormat("#.0");
+        greenTerraLength = String.valueOf(2 + ((double) skillValue / (double) abilityLengthIncreaseLevel));
+        //FARMERS DIET
+        if(skillValue >= farmersDietMaxLevel) farmersDietRank = "5";
+        else farmersDietRank = String.valueOf((double) skillValue / (double) farmersDietRankChange);
+        //GREEN THUMB
+        if(skillValue >= greenThumbStageMaxLevel) greenThumbStage = "4";
+        else greenThumbStage = String.valueOf((double) skillValue / (double) greenThumbStageChange);
+        
+        
+        if(skillValue >= greenThumbMaxLevel) greenThumbChance = String.valueOf(greenThumbMaxBonus);
+        else greenThumbChance = String.valueOf(((double) greenThumbMaxBonus / (double) greenThumbMaxLevel) * (double) skillValue);
+        //DOUBLE DROPS
+        if(skillValue >= doubleDropsMaxLevel) doubleDropChance = df.format(doubleDropsMaxBonus);
+        else doubleDropChance = df.format(((double) doubleDropsMaxBonus / (double) doubleDropsMaxLevel) * (double) skillValue);
     }
 
     @Override
